+++ lib/raw_syntax/raw_term.ml
@@ -21,12 +21,6 @@ let call (op, args) = Call (Symbol.of_string op, args)
 
 let var_list symbols = List.map (fun x -> Var x) symbols
 
-let is_immediate = function
-  | Var _ | Const _ -> true
-  | Call (c, []) when Symbol.op_kind c = `CCall -> true
-  | Call _ | Match _ | Let _ -> false
-;;
-
 let to_string =
     let rec go = function
       | Var x -> Symbol.to_string x